requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
certifi==2024.8.30
coverage==7.6.8
distro==1.9.0
Django==5.1.2
django-filter==24.3
djangorestframework==3.15.2
drf-yasg==1.21.8
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
jiter==0.8.0
openai==1.55.3
packaging==24.2
pillow==11.0.0
pluggy==1.5.0
pydantic==2.10.2
pydantic_core==2.27.1
pytest==8.3.4
pytest-django==4.9.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
sniffio==1.3.1
sqlparse==0.5.1
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
